Simplify exception handling logic in lib/upload.py
Notably, remove the use of boolean return values and rely exclusively
on exception hierarchy and retry-able vs. non-retryable exceptions.
This includes minor flow control logic changes (exceptions vs. return
values), but no major changes to the upload code paths.
